task: fix uart
I should further note that this is an absolutely rediculous
way to build a system. I have not yet in my academic life
found a system this poorly documented and seemingly opaque.

This experience alone may force me to reevaluate my
problem list. I think the embedded ecosystem needs a great
deal of work.

If it did not come through in this text, I would like to add
for the record that I am utterly incensed. Truly furious.
Arguably the _most_ angry I have been in years. I truly have
not felt the blood roll to my head like this in many years.